+++ src/utf8_sfml_redefinitions.hpp
@@ -3,6 +3,7 @@
 #include <SFML/Audio.hpp>
 #include <SFML/Audio/InputSoundFile.hpp>
 #include <SFML/Audio/SoundBuffer.hpp>
+#include <SFML/Graphics/Shader.hpp>
 #include <SFML/Graphics/Texture.hpp>
 
 #include "utf8_sfml.hpp"
@@ -12,4 +13,5 @@ namespace feis {
     using InputSoundFile = feis::UTF8Streamer<sf::InputSoundFile>;
     using SoundBuffer = feis::UTF8Loader<sf::SoundBuffer>;
     using Texture = feis::UTF8Loader<sf::Texture>;
+    using Shader = feis::UTF8Loader<sf::Shader>;
 }